Move core component from zope.app.session to zope.session
Note:
Some of the next commits will be broken because I have a 
chicken and egg problem.

The zope.session uses zope.app.authentication in it's test
setup. This requires to change the zope.app.session, zope.session
and zope.app.authentication all at once which is not possible.
Because they depend in it's test setup on each other.

I'll do the next steps in the nice old trunk, it's impossible to 
do it with the egg development process. So I will add eggs later
after the trunk works