Add support for PyPy3.
PyPy3 is a bit of an add duck, supporting the 3.3 syntax (u"") but providing the 3.2 library. When I had the library issue fixed, I fixed the syntax issue and tried to also add Python 3.2, but our dependency on zope.publisher (at least) requires at least 3.3 because of syntax. So if you want to run this under 3.2, PyPy3 is your only option.